:art: Make components reusable
