+++ src/components/SearchBox.tsx
@@ -1,7 +1,6 @@
 import { useRef, useState } from "react";
 import { useAppDispatch } from "../redux/configStore.hooks";
 import { setSearchResult, setQuery } from "../redux/features/searchSlice";
-import searchImg from "/src/assets/search-20x20.png";
 
 const SearchBox = (props: { placeholder: string }) => {
   const dispatch = useAppDispatch();
@@ -28,17 +27,14 @@ const SearchBox = (props: { placeholder: string }) => {
   const queryRef = useRef<HTMLInputElement>(null);
 
   return (
-    <div className="hbox w(370) h(40) b(2/#505050) r(20) pl(12)">
-      <img src={searchImg} className="w(20) h(20) mr(12)" alt="search" />
-      <input
-        onChange={onQueryChange}
-        value={inputValue}
-        ref={queryRef}
-        spellCheck={false}
-        className="placeholder::font(14)+900+c(#aaa) NanumGothic font(14) 900 c(#363636) dark-white w(300) dark-transparent"
-        placeholder={props.placeholder}
-      />
-    </div>
+    <input
+      onChange={onQueryChange}
+      value={inputValue}
+      ref={queryRef}
+      spellCheck={false}
+      className="placeholder::font(14)+900+c(#aaa) NanumGothic font(14) 900 c(#363636) dark-white w(300) dark-transparent"
+      placeholder={props.placeholder}
+    />
   );
 };
